+++ src/index.tsx
@@ -1,5 +1,6 @@
 export { MalvoConnect } from "./MalvoConnect";
 export { buildConnectUrl, buildFinishUrl } from "./url";
+export { openPaymentAuthorization } from "./payment";
 export type {
   MalvoConnectOptions,
   MalvoSuccess,